# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: VOID - Multi-artist link-in-bio platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /me:
    get:
      operationId: getMe
      tags: [profile]
      summary: Get current authenticated user profile
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "401":
          description: Unauthorized
    put:
      operationId: updateProfile
      tags: [profile]
      summary: Update authenticated user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "401":
          description: Unauthorized
  /me/username-check:
    get:
      operationId: checkUsernameAvailability
      tags: [profile]
      summary: Check if a username is available
      parameters:
        - name: username
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Availability result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsernameCheckResult"
  /profile/{username}:
    get:
      operationId: getPublicProfile
      tags: [profile]
      summary: Get public artist profile with links and photos
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public artist profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicProfile"
        "404":
          description: Profile not found
...
  /photos/public/{username}:
    get:
      operationId: getPublicPhotos
      tags: [photos]
      summary: Get all public photos for an artist
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Artist photos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Photo"
        "404":
          description: Profile not found
  /photos:
    get:
      operationId: getMyPhotos
      tags: [photos]
      summary: Get all photos for the authenticated user
      responses:
        "200":
          description: User photos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Photo"
        "401":
          description: Unauthorized
    post:
      operationId: createPhoto
      tags: [photos]
      summary: Save photo metadata after upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PhotoInput"
      responses:
        "201":
          description: Created photo record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Photo"
        "401":
          description: Unauthorized
  /photos/reorder:
    put:
      operationId: reorderPhotos
      tags: [photos]
      summary: Reorder photos by updating positions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderInput"
      responses:
        "200":
          description: Reordered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "401":
          description: Unauthorized
  /photos/{id}:
    put:
      operationId: updatePhoto
      tags: [photos]
      summary: Update photo metadata
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PhotoUpdate"
      responses:
        "200":
          description: Updated photo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Photo"
        "401":
          description: Unauthorized
        "404":
          description: Photo not found
    delete:
      operationId: deletePhoto
      tags: [photos]
      summary: Delete a photo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "401":
          description: Unauthorized
        "404":
          description: Photo not found
...
components:
  schemas:
...
    Profile:
      type: object
      required: [id, clerkUserId, username, isActive, createdAt]
      properties:
        id:
          type: string
        clerkUserId:
          type: string
        username:
          type: string
        displayName:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        headerImageUrl:
          type: ["string", "null"]
        accentColor:
          type: ["string", "null"]
        bgColor:
          type: ["string", "null"]
        cardStyle:
          type: ["string", "null"]
        isSuperAdmin:
          type: ["boolean", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
    ProfileUpdate:
      type: object
      properties:
        username:
          type: string
        displayName:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
        headerImageUrl:
          type: string
        accentColor:
          type: string
        bgColor:
          type: string
        cardStyle:
          type: string
    PublicProfile:
      type: object
      required: [profile, links, photos]
      properties:
        profile:
          $ref: "#/components/schemas/Profile"
        links:
          type: array
          items:
            $ref: "#/components/schemas/Link"
        photos:
          type: array
          items:
            $ref: "#/components/schemas/Photo"
        socialLinks:
          type: array
          items:
            $ref: "#/components/schemas/SocialLink"
    UsernameCheckResult:
      type: object
      required: [available]
      properties:
        available:
          type: boolean
    Link:
      type: object
      required: [id, profileId, title, url, position, isVisible, clickCount, createdAt]
      properties:
        id:
          type: string
        profileId:
          type: string
        title:
          type: string
        url:
          type: string
        description:
          type: ["string", "null"]
        icon:
          type: ["string", "null"]
        thumbnailUrl:
          type: ["string", "null"]
        cardType:
          type: string
        position:
          type: integer
        isVisible:
          type: boolean
        clickCount:
          type: integer
        createdAt:
          type: string
...
    Photo:
      type: object
      required: [id, profileId, url, isCover, position, createdAt]
      properties:
        id:
          type: string
        profileId:
          type: string
        url:
          type: string
        caption:
          type: ["string", "null"]
        isCover:
          type: boolean
        position:
          type: integer
        createdAt:
          type: string
    PhotoInput:
      type: object
      required: [url]
      properties:
        url:
          type: string
        caption:
          type: string
        isCover:
          type: boolean
    PhotoUpdate:
      type: object
      properties:
        caption:
          type: string
        isCover:
          type: boolean
        position:
          type: integer
    SocialLink:
      type: object
      required: [id, profileId, platform, url, position]
      properties:
        id:
          type: string
        profileId:
          type: string
        platform:
          type: string
        url:
          type: string
        position:
          type: integer
    ReorderInput:
      type: object
      required: [ids]
      properties:
        ids:
          type: array
          items:
            type: string
...
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
...