# lib/api-zod/src/generated/types/profileUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VOID - Multi-artist link-in-bio platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ProfileUpdate {
  username?: string;
  displayName?: string;
  bio?: string;
  avatarUrl?: string;
  headerImageUrl?: string;
  accentColor?: string;
  bgColor?: string;
  cardStyle?: string;
}