# lib/api-zod/src/generated/types/socialLink.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VOID - Multi-artist link-in-bio platform API
 * OpenAPI spec version: 0.1.0
 */

export interface SocialLink {
  id: string;
  profileId: string;
  platform: string;
  url: string;
  position: number;
}